create Person module
